requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
groq==0.33.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.4
openai==2.6.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
psutil==7.1.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0


# Core ML dependencies
torch>=2.0.0,<3.0.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# ONNX and optimization - FIXED
onnx>=1.14.0
onnxruntime>=1.16.0
optimum[onnxruntime]>=1.14.0
onnx-graphsurgeon>=0.3.27
onnxscript>=0.1.0
# Alternative if above doesn't work:
# optimum>=1.14.0
# onnxruntime>=1.16.0

# FastAPI and serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Utilities and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
jupyter>=1.0.0

matplotlib>=3.5.0
seaborn>=0.11.0

wandb
google-genai
ijson>=3.1.4
tabulate>=0.9.0

# Optional: For advanced optimization
# intel-extension-for-pytorch  # Uncomment for Intel CPU optimization
# mkl  # Uncomment for Intel MKL support
